+++ src/app/locales/en/components/modals/record-list-selector-modal.js
@@ -1,0 +1,5 @@
+export default {
+  nothingToShow: 'There are no records to show.',
+  cancelBtn: 'Cancel',
+  confirmBtn: 'Confirm selection',
+};