+++ src/app/locales/en/components/space-automation/input-stores-form.js
@@ -4,4 +4,13 @@ export default {
     buttonText: 'Use selected items ({{itemsCount}})',
   },
   itemsSelectorSubheader: 'for <strong>{{storeName}}</strong> store',
+  groupSelector: {
+    header: {
+      single: 'Select a group',
+      multi: 'Select groups',
+    },
+    listHeader: 'Groups',
+    incompleteListText: 'Some space groups may not be visible due to limited privileges.',
+    incompleteListTipText: 'Since you do not have the <strong>view space</strong> privilege in this space, you cannot view its full member list. You can however see the member groups that you belong to.',
+  },
 };