+++ src/app/locales/en/components/visual-edm/object.js
@@ -2,8 +2,8 @@ import EdmObjectType from 'oneprovider-gui/utils/edm/object-type';
 
 export const objectTypeName = Object.freeze({
   [EdmObjectType.Aggregation]: 'Aggregation',
-  [EdmObjectType.ProvidedCHO]: 'Provided Cultural Heritage Object',
-  [EdmObjectType.WebResource]: 'Web Resource',
+  [EdmObjectType.ProvidedCHO]: 'Cultural Heritage Object',
+  [EdmObjectType.WebResource]: 'Digital Object',
 });
 
 export default {
@@ -16,9 +16,11 @@ export default {
   notValid: 'Definition of this {{objectType}} is not valid',
   objectTypeTip: {
     [EdmObjectType.Aggregation]: 'Aggregated information about all related resources pertaining to the provided CHO',
-    [EdmObjectType.ProvidedCHO]: 'Information related to the original Cultural Heritage Object (CHO)',
+    [
+      EdmObjectType.ProvidedCHO
+    ]: 'Information related to the original Cultural Heritage Object (CHO) — refers to the ProvidedCHO of EDM',
     [
       EdmObjectType.WebResource
-    ]: 'Information related to a digital representation of the object (there may be more than one)',
+    ]: 'Information related to a digital representation of the object — refers to the WebResource of EDM',
   },
 };